feat: Add CSS variables for light and dark themes
This commit introduces CSS variables for light and dark themes. The
variables --light-bg and --dark-bg are added to the :root selector.
These variables are then used to set the background color of the body
and the color of the link in the #githublink element. This change
allows for easier theme management and potential future theme
expansions.